Use a trait for approximate equality
This is instead of implementing the existing trait `PartialEq`. I'm
still learning Rust, so this is unfamiliar and possibly wrong, but I
think that I'm not permitted to provide a `PartialEq` implementation to
a type that isn't "mine"--but, I *can* provide an implementation of my
_own_ trait.

I think. I'm still playing around with this.